+++ R/hr_overlap.R
@@ -1,8 +1,14 @@
 #' Different Methods to calculate home-range overlaps
 #'
 #' @param x,y `hr` A home-range estimate
+#' @param type `[character](1)` \cr Type of index, should be one of `hr`, `phr`,
+#'   `vi`, `ba`, `udoi`, or `hd`.
+#' @param conditional `[logical](1)` \cr Whether or not conditional UDs are
+#'   used. If `TRUE` levels from that were used to estimate home ranges will be
+#'   used.
 #' @template dots_none
-#' @return \code{data.frame} with the isopleth level and area in units of the coordinate reference system.
+#' @return \code{data.frame} with the isopleth level and area in units of the
+#'   coordinate reference system.
 #' @name hr_overlaps
 NULL
 
@@ -19,22 +25,108 @@ hr_overlap.default <- function (x, ...) {
 
 #' @export
 #' @rdname hr_overlaps
-hr_overlap.hr <- function(x, y, ...) {
+hr_overlap.hr <- function(x, y, type = "hr", conditional = FALSE, ...) {
 
-  if (!inherits(y, "hr")) {
-    stop("y is no home-range estimate.")
+  checkmate::assert_class(x, "hr")
+  checkmate::assert_class(y, "hr")
+  checkmate::assert_character(type, len = 1)
+  checkmate::assert_logical(conditional, len = 1)
+
+  if (!type %in% c("hr", "phr", "vi", "ba", "udoi", "hd")) {
+    stop("type should be one of: hr, phr, vi, ba, udoi, or hd")
   }
-  x <- hr_isopleths(x)
-  y <- hr_isopleths(y)
- overlap_base(x, y)
+
+  if (type == "hr") {
+    x <- hr_isopleths(x)
+    y <- hr_isopleths(y)
+    hr_base(x, y)
+  } else {
+    # both of correct class
+    if (is(x, "hr_prob") & is(y, "hr_prob")) {
+      xud <- hr_ud(x)
+      yud <- hr_ud(y)
+
+      # Check that both are of the same raster extent
+      if (identical(raster::extent(xud), raster::extent(yud)) &
+          all(raster::res(xud) == raster::res(yud))) {
+
+        # Do I have to get conditional uds?
+        if (conditional) {
+          # Check x and y have the same levels
+          if (!all(x$levels == y$levels)) {
+            stop("x and y should have the same level for conditional hr_overlap")
+          }
+          tibble(
+            levels = x$levels,
+            overlap = purrr::map_dbl(levels, ~ vol_base(
+              get_cond_ud(x, level = .x)[],
+              get_cond_ud(y, level = .x)[],
+              type = type
+            ))
+          )
+        } else {
+          tibble(
+            levels = 1,
+            overlap = vol_base(xud[], yud[], type)
+          )
+        }
+      }
+    } else {
+      stop("Both x and y should be probabilistic estimators.")
+    }
+  }
+}
+
+get_cond_ud <- function(x, level = 1) {
+    cud <- hr_cud(x) # This is the cumulative UD (i.e., something like the CDF)
+    ud <- hr_ud(x)
+    ud[cud > level] <- 0
+    ud[] <- ud[] / level
+    ud
+}
+
+hr_base <- function(x, y) {
+  if(all(x$levels == y$levels)) {
+    if (length(x$level) == 1) {
+      return(
+        tibble::tibble(
+          level = x$level,
+          overlap = hr_base_ov(x, y))
+      )
+    } else if (length(x$level) > 1) {
+      return(
+        tibble::tibble(
+          levels = x$level,
+          overlap = sapply(1:nrow(x), function(i) hr_base_ov(x[i, ], y[i, ])))
+      )
+    }
+  } else {
+    stop("Not all levels of x in y.")
+  }
+}
+
+vol_base <- function(ud_i, ud_j, type) {
+  if (type == "phr") {
+    sum(ud_j[ud_i > 0])
+  } else if (type == "vi") {
+    sum(pmin(ud_i, ud_j))
+  } else if (type == "ba") {
+    sum(sqrt(ud_i) * sqrt(ud_j))
+  } else if (type == "udoi") {
+    A_ij <- sum(ud_i > 0 & ud_j > 0)
+    A_ij * sum(ud_i * ud_j)
+  } else if (type == "hd") {
+    2 * (1 - sum(sqrt(ud_i) * sqrt(ud_j)))
+  }
 }
 
 
 #' @rdname hr_overlaps
 #' @export
-#' @param  consecutive.only `[logical=TRUE]` \cr Should only consecutive overlaps be calculated or all combinations?
+#' @param  which `[character = "consecutive"]` \cr Should only consecutive overlaps be calculated or all combinations?
 #' @param  labels `[character=NULL]` \cr Labels for different instances. If `NULL` (the default) numbers will be used.
-hr_overlap.list <- function(x, consecutive.only = TRUE, labels = NULL, ...) {
+hr_overlap.list <- function(
+  x, type = "hr", conditional = FALSE,  which = "consecutive", labels = NULL, ...) {
 
   # check all elements are hr
   if(!all(sapply(x, inherits, "hr"))) {
@@ -45,14 +137,23 @@ hr_overlap.list <- function(x, consecutive.only = TRUE, labels = NULL, ...) {
     stop("At least two home range estimates are needed")
   }
 
+  checkmate::assert_character(which)
+  if (!which %in% c("all", "consecutive", "one_to_all")) {
+    stop("Which should be one of 'all', 'consecutive' or 'one_to_all'")
+  }
+
   nn <- if (is.null(names(x))) 1:length(x) else names(x)
 
   if (!is.null(labels)) {
+    if (!is.character(labels)) {
+      labels <- as.character(labels)
+      message("labels coerced to character.")
+    }
     if (length(labels) == length(nn)) {
       nn <- labels
+      names(x) <- labels
     } else {
       warning("Length of `labels` does not match the number of cases. Using default instead.")
-
     }
   }
 
@@ -64,99 +165,91 @@ hr_overlap.list <- function(x, consecutive.only = TRUE, labels = NULL, ...) {
     stop("Duplicated names are not permitted")
   }
 
-  isos <- lapply(x, hr_isopleths)
-  levels <- lapply(isos, function(x) x$level)
-  if (!all(sapply(levels[-1], function(x) all(levels[[1]] %in% x)))) {
+  levels <- lapply(x, function(y) y$levels)
+  if (!all(sapply(levels[-1], function(x) all(levels[[1]] == x)))) {
     stop("Not all levels of first home range est are in the others as well.")
   }
 
-  grid <- if (consecutive.only) {
+  grid <- if (which == "consecutive") {
     tibble(from = nn[-length(nn)], to = nn[-1])
-  } else {
+  } else if (which == "all") {
     tidyr::expand_grid(from = nn, to = nn) %>%
       dplyr::filter(from != to)
+  } else if (which == "one_to_all") {
+    tibble(from = nn[1], to = nn[-1])
   }
+
 
   grid %>%
     dplyr::mutate(overlap = purrr::map2(from, to, function(i, j) {
-      overlap_base(isos[[i]], isos[[j]])
+      hr_overlap(x[[i]], x[[j]], type = type, conditional = conditional)
     })) %>%
     tidyr::unnest(cols = "overlap")
 }
 
-overlap_base <- function(x, y) {
 
-  if(all(x$levels %in% y$levels)) {
-    if (length(x$level) == 1) {
-      if (sf::st_intersects(x, y, sparse = FALSE)) {
-        suppressWarnings(ol <- sf::st_intersection(x, y))
-        return(as.numeric(sf::st_area(ol) / sf::st_area(x)))
-      } else {
-        return(0)
-      }
-    } else if (length(x$level) > 1) {
-      return(
-        tibble::tibble(
-          level = x$level,
-          overlap = sapply(1:nrow(x), function(i) overlap_base(x[i, ], y[i, ])))
-        )
-    }
-  } else {
-    stop("Not all levels of x in y.")
-  }
-}
-
-#' Different Methods to calculate home-range intersections
+#' Calculate the overlap between a home-range estimate and a polygon
 #'
-#' @param x,y A probabilistic home-range estimate.
-#' @param type `[character]` \cr Which index should be calculated.
-#' @param conditional `[numeric]` \cr Condition on which number?
-#' @template dots_none
-#' @return \code{data.frame} with the isopleth level and area in units of the coordinate reference system.
-#' @name hr_intersection
+#' @param x A home-range estimate.
+#' @param sf An object of class `sf` containing polygons
+#' @param direction The direction.
+#' @param feature_names  optional feature names
+#'
+#' @return A tibble
 #' @export
-hr_intersection <- function (x, ...) {
-  UseMethod ("hr_intersection", x )
+#'
+#' @examples
+hr_overlap_feature <- function(x, sf, direction = "hr_with_feature", feature_names = NULL) {
+
+  checkmate::assert_class(x, "hr")
+  checkmate::assert_class(sf, "sf")
+  checkmate::assert_character(direction, len = 1)
+
+  if (!sf::st_geometry_type(sf) %in% c("POLYGON", "MULTIPOLYGON")) {
+    stop("`sf` should be a (MULTI)POLYGON.")
+  }
+
+  if (!direction %in% c("hr_with_feature", "feature_with_hr")) {
+    stop("Wrong direction")
+  }
+  if (!is.null(feature_names)) {
+    if (length(feature_names) != nrow(sf)) {
+      stop("`feature_names` do not match the number of features in `sf`.")
+    }
+  }
+  names_hr <- x$levels
+  names_sf <- if(is.null(feature_names)) 1:nrow(sf) else feature_names
+
+  if (direction == "hr_with_feature") {
+    hr_feature(hr_isopleths(x), sf) %>%
+      mutate(from = names_hr[from], to = names_sf[to])
+  } else if (direction == "feature_with_hr") {
+    hr_feature(sf, hr_isopleths(x)) %>%
+      mutate(from = names_sf[from], to = names_hr[to])
+  }
 }
 
-#' @export
-#' @rdname hr_intersection
-hr_intersection.hr_prob <- function(x, y, type = "ba", conditional = 0.95, ...) {
 
-  if (!is(x, "hr_prob") & is(y, "hr_prob")) {
-    stop("x, y need to be prob estimators.")
+hr_feature <- function(x, y) {
+  if (nrow(x) == 1 & nrow(y) == 1) {
+    return(hr_base_ov(x, y))
+  } else {
+    res <- expand.grid(from = 1:nrow(x), to = 1:nrow(y))
+    return(
+      tibble::tibble(
+        from = res$from,
+        to = res$to,
+        overlap = purrr::map_dbl(1:nrow(res), function(i) hr_base_ov(x[res$from[i], ], y[res$to[i], ]))
+      )
+    )
   }
-
-  x_ud <- hr_ud(x)
-  y_ud <- hr_ud(y)
-
-  if (conditional < 1) {
-    x_ud[] <- ifelse(hr_cud(x)[] <= conditional, x_ud[], 0) / conditional
-    y_ud[] <- ifelse(hr_cud(y)[] <= conditional, y_ud[], 0) / conditional
-  }
-
 }
 
-hr_intersection_base <- function(x, y, type) {
-
-  if (!identical(raster::extent(x), raster::extent(y))) {
-    stop("x and y do not have an identical extent")
-  }
-  r1 <- x[]
-  r2 <- y[]
-  r1 <- r1 / sum(r1)
-  r2 <- r2 / sum(r2)
-
-  if (type == "phr") {
-    sum(r2[r1 > 0])
-  } else if (type == "ba") {
-    ## bhattacharyya's afinity
-    sum(sqrt(r1) * sqrt(r2))
-  } else if (type == "vi") {
-    sum(pmin(r1, r2))
-  } else if (type == "hd") {
-    2 * (1 - sum(sqrt(r1) * sqrt(r2)))
-  } else if (type == "udoi") {
-    NA
+hr_base_ov <- function(x, y) {
+  if (sf::st_intersects(x, y, sparse = FALSE)) {
+    suppressWarnings(ol <- sf::st_intersection(x, y))
+    as.numeric(sf::st_area(ol) / sf::st_area(x))
+  } else {
+    0
   }
 }