+++ R/track.R
@@ -212,13 +212,18 @@ nest.track_xy <- function(.data, ..., .dots) {
 }
 
 #' @export
-unnest.nested_track <- function(.data, cols, ...) {
+unnest.nested_track <- function(
+  data, cols, ...
+  #keep_empty = FALSE, ptype = NULL, names_sep = NULL,
+  ####names_repair = "check_unique",
+  ############.drop = NULL, .id = NULL, .sep = NULL, .preserve = NULL
+  ) {
   col <- rlang::as_label(rlang::enquo(cols))
   if (length(col) > 1) {
     stop("amt currently only supports unnesting one column")
   }
 
-  dd <- .data[[col]]
+  dd <- data[[col]]
 
   # check length > 1
   class_col_entries <- if (length(dd) > 1) {
@@ -234,8 +239,12 @@ unnest.nested_track <- function(.data, cols, ...) {
     }
   }
   # check they are all of the same class
-  class(.data) <- c("tbl_df", "tbl", "data.frame")
-  x <- unnest(.data, cols = !!rlang::enquo(cols))
+  class(data) <- c("tbl_df", "tbl", "data.frame")
+  x <- unnest(data, cols = !!rlang::enquo(cols))
+ #             keep_empty = keep_empty, ptype = ptype,
+ ####             names_sep = names_sep, names_repair = names_repair,
+ #             .drop = .drop, .id = .id, .sep = .sep,
+ ####             .preserve = .preserve)
   class(x) <- class_col_entries
   track_transfer_attr(dd[[1]], x)
 }