+++ NEWS.md
@@ -2,7 +2,11 @@
 ## updates
 - `unnest` works for track_xy*
 - Major updates for `hr_akde`. This includes that now CI are available for  
--
+- new vignette and methods for `hr_overlap`
+- reduced package dependencies.
+
+## bug fixes
+- Fixed bug in confidence calculation for `log_rss`.
 
 
 # amt 0.1.3